file and function renaming so that it is always the same

+++ benchmarks/dart/test/Eq/oldV.c
@@ -1,6 +1,4 @@
-#include <stdio.h>
-#include <math.h>
-double snippet(int x, int y) {
+int snippet(int x, int y) {
         if (x*x*x > 0){
             if(x>0 && y==10)
                 return 1000;
